World File
We create a correlation matrix to view the correlation coefficient between each pairwise combination of numeric variables in the dateset
 We estimate the parameters of the model and the value of the y-intercept (in this case 71.72464) and the estimated effect of body fat on abdomen (1.07440).p value (here it is 2.2e-16, or almost zero), which will indicate whether the model fits the data well.From these results we can reject null hypothesis and accept alternative hypothesis we can say that there is a significant positive relationship between body fat and abdomen. Residual standard error (RSE), R-squared (R2) and the F-statistic are metrics that are used to check how well the model fits to our data.They give overall quality of the model

evidence as to whether assumptions of the model have been met;
 
conduct a formal test to question whether there is a significant linear relationship between body fat and abdomen size.Summarizing a dateset using descriptive statistics. Visualizing a dateset using charts. Identifying missing values. We create a correlation matrix to view the correlation coefficient between each pairwise combination of numeric variables in the dateset We estimate the parameters of the model and the value of the y-intercept (in this case 71.72464) and the estimated effect of body fat on abdomen (1.07440).p value (here it is 2.2e-16, or almost zero), which will indicate whether the model fits the data well.From these results we can reject null hypothesis and accept alternative hypothesis we can say that there is a significant positive relationship between body fat and abdomen. Residual standard error (RSE), R-squared (R2) and the F-statistic are metrics that are used to check how well the model fits to our data.They give overall quality of the model evidence as to whether assumptions of the model have been met; conduct a formal test to question whether there is a significant linear relationship between body fat and abdomen size.
Skills: Analytical Skills · R (Programming Language) · Hypothesis Testing · Correlation Analysis · Statistical Data Analysis · Data Visualization